requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
Click==7.0
cycler==0.10.0
Cython==0.29.7
Flask==1.0.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
imutils==0.5.2
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
numpy==1.16.3
opencv-contrib-python==4.1.0.25
pbr==5.2.0
Pillow==6.0.0
protobuf==3.7.1
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.9.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
